CFormat single pass for color matching and skipping over
This new way of checking and skipping colors relies on the fact that a color tag should be correct, so is a rather naive operation, this is being done by returning the length of the color tag.
We are also now not copying the color tag to a color buffer as we can now set the delimiter for the color map.